Fix: ランダムにピックアップされた数字も必要になる為修正
ランダムにピックアップされた文字はランダムな数字から抽出するようにする
