+++ app/javascript/store/index.js
@@ -1,13 +1,13 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 import selectedWords from './modules/selectedWords'
-import resultWord from './modules/resultWord'
+import randomPickedUp from './modules/randomPickedUp'
 
 Vue.use(Vuex)
 
 export default new Vuex.Store({
   modules: {
     selectedWords,
-    resultWord
+    randomPickedUp
   }
 })
